功能(出库)：wave4 仓储 ERP 取消与数量 Decimal 化
ERP 取消命令与 PendingErpCancel 守卫；数量字段 Decimal 化并
对齐 NUMERIC(19,4)。

关联：H8 v1.9

+++ backend/crates/api/src/wave4_repository_waves.rs
@@ -128,7 +128,7 @@ impl PgWave4Repository {
                SET qty_locked = batch.qty_locked - released.qty,
                    updated_at = $3
               FROM (
-                    SELECT allocation.batch_id, SUM(allocation.allocated_qty)::BIGINT AS qty
+                    SELECT allocation.batch_id, SUM(allocation.allocated_qty) AS qty
                       FROM inventory_allocations allocation
                       JOIN outbound_wave_orders link
                         ON link.owner_id = allocation.owner_id
